feat: 좋아요 커스텀 훅 테스트
